[MOD] project 구조 변경
